fix(DiskSource): fix Uint8 overflow
Replace Uint8Array by Uint32Array to fix overflow error with
disks with more than 50 points

+++ Sources/Filters/Sources/DiskSource/index.js
@@ -87,7 +87,7 @@ function vtkDiskSource(publicAPI, model) {
 
     // Generate cell connectivity (quads)
     const cellCount = radialResolution * circumferentialResolution;
-    const cellData = new Uint8Array(cellCount * 5);
+    const cellData = new Uint32Array(cellCount * 5);
     let offset = 0;
     for (let i = 0; i < model.circumferentialResolution; i++) {
       for (let j = 0; j < model.radialResolution; j++) {
